exceptions added to methods!

+++ src/main/java/al/ikubinfo/internship/freelancer/service/impl/ExperienceServiceImple.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import al.ikubinfo.internship.freelancer.entity.Experience;
+import al.ikubinfo.internship.freelancer.exception.ResourceNotFoundException;
 import al.ikubinfo.internship.freelancer.mapper.Mapper;
 import al.ikubinfo.internship.freelancer.model.ExperienceModel;
 import al.ikubinfo.internship.freelancer.repository.ExperienceRepository;
@@ -30,8 +31,26 @@ public class ExperienceServiceImple implements ExperienceService {
 	}
 
 	@Override
-	public ExperienceModel addOrUpdate(ExperienceModel experienceModel) {
+	public ExperienceModel addExperience(ExperienceModel experienceModel) {
 		Experience expEntity = expMapper.toEntity(experienceModel);
+		repository.save(expEntity);
+		return expMapper.toModel(expEntity);
+	}
+
+	@Override
+	public ExperienceModel updateExperience(Integer id, ExperienceModel experienceModel) {
+		Experience expEntity = expMapper.toEntity(experienceModel);
+		Experience expById = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Not found experience with id" + id));
+
+		expEntity.setId(expById.getId());
+		expEntity.setNameOfCompany(expById.getNameOfCompany());
+		expEntity.setPosition(expById.getPosition());
+		expEntity.setPositionDescription(expById.getPositionDescription());
+		expEntity.setStartDate(expById.getStartDate());
+		expEntity.setEndDate(expById.getEndDate());
+		expEntity.setUser(expById.getUser());
+
 		repository.save(expEntity);
 		return expMapper.toModel(expEntity);
 	}
@@ -46,17 +65,24 @@ public class ExperienceServiceImple implements ExperienceService {
 	@Override
 	public String deleteExperience(int id) {
 		Experience experience = repository.getOne(id);
+		if(experience==null) {
+			throw new ResourceNotFoundException("enter a valid experience id");
+		}
 		repository.delete(experience);
 		return String.format("Experience with %d , position %s at company %s IS DELETED", id, experience.getPosition(),
 				experience.getNameOfCompany());
 
-	}//handle exception if an id does not exists
+	}
 
 	@Override
 	public List<ExperienceModel> getExperiencesByUserId(int userId) {
+
 		List<Experience> experienceEntityList = repository.findByUserId(userId);
+		if (experienceEntityList.isEmpty()) {
+			throw new ResourceNotFoundException("NO CONTENT");
+		}
+
 		return expMapper.toModelList(experienceEntityList);
 	}
 
-	
 }
